% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{elu_R}
\alias{elu_R}
\title{Elu Activation Function in R}
\usage{
elu_R(x, alpha = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{A numeric scalar.}
}
\description{
Elu Activation Function in R
}
