% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{exponential_R}
\alias{exponential_R}
\title{Exponential Activation Function in R}
\usage{
exponential_R(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Exponential Activation Function in R
}
