% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{hard_sigmoid_R}
\alias{hard_sigmoid_R}
\title{Hard Sigmoid Activation Function in R}
\usage{
hard_sigmoid_R(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Hard Sigmoid Activation Function in R
}
