% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{relu_R}
\alias{relu_R}
\title{Relu Activation Function in R}
\usage{
relu_R(x, alpha = 0, max_value = Inf, threshold = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{A numeric scalar.}

\item{max_value}{A numeric scalar.}

\item{threshold}{A numeric scalar.}
}
\description{
Relu Activation Function in R
}
