% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_regression_data.R
\name{simulate_regression_data}
\alias{simulate_regression_data}
\title{Simulate data for a Neural Network Structure.}
\usage{
simulate_regression_data(rows = 1000, N = 5, U = 5, C = 0,
  matrices, activations, noise = 0)
}
\arguments{
\item{rows}{An integer scaler. The number of rows in simulated data.}

\item{N}{An integer scaler. The number of normal random variables in X.}

\item{U}{An integer scaler. The number of uniform random variables in X.}

\item{C}{An integer scaler. The number of binary random variables in X.}

\item{matrices}{A list. Each element is a matrix defining the structure the layer.}

\item{activations}{A list. Each element is a activation function.}

\item{noise}{A positive numeric scaler.
The standard deviation of a normal random variable added to response.
0 is no noise.}
}
\value{
A matrix containing predictor variables and a response variable.
}
\description{
Simulate data for a Neural Network Structure.
}
\examples{
library(NeuralNetworkSimulatoR)

# Linear regression with weights 1, 2, and 3
M <- list(matrix(1:3, nrow = 3, ncol = 1))
A <- list(linear_R)
simData <- simulate_regression_data(
  rows = 1000L,
  N = 3L, U = 0L, C = 0L,
  matrices = M, activations = A,
  noise = 0
)
rm(A, M)

# Complex network
# 10 nodes in first hidden layer. Activation relu
# 5 nodes in second hidden layer  Activation elu

M <- list(matrix(1:10, nrow = 10, ncol = 5), matrix(1:5, nrow = 5, ncol = 1))
A <- list(relu_R, elu_R)
simData <- simulate_regression_data(
  rows = 1000L,
  N = 5L, U = 5L, C = 0L,
  matrices = M, activations = A,
  noise = 0
)
rm(A, M)

rm(simData)
}
